+++ gen/messenger_grpc.pb.go
@@ -101,6 +101,7 @@ const (
 	ChatService_SubscribeNotifications_FullMethodName = "/messenger.ChatService/SubscribeNotifications"
 	ChatService_GetNotificationHistory_FullMethodName = "/messenger.ChatService/GetNotificationHistory"
 	ChatService_MarkNotificationsRead_FullMethodName  = "/messenger.ChatService/MarkNotificationsRead"
+	ChatService_GetUnreadCount_FullMethodName         = "/messenger.ChatService/GetUnreadCount"
 )
 
 // ChatServiceClient is the client API for ChatService service.
@@ -195,6 +196,7 @@ type ChatServiceClient interface {
 	SubscribeNotifications(ctx context.Context, in *SubscribeNotificationsRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[ServerNotification], error)
 	GetNotificationHistory(ctx context.Context, in *GetNotificationHistoryRequest, opts ...grpc.CallOption) (*GetNotificationHistoryResponse, error)
 	MarkNotificationsRead(ctx context.Context, in *MarkNotificationReadRequest, opts ...grpc.CallOption) (*MarkNotificationReadResponse, error)
+	GetUnreadCount(ctx context.Context, in *GetUnreadCountRequest, opts ...grpc.CallOption) (*GetUnreadCountResponse, error)
 }
 
 type chatServiceClient struct {
@@ -1070,6 +1072,16 @@ func (c *chatServiceClient) MarkNotificationsRead(ctx context.Context, in *MarkN
 	return out, nil
 }
 
+func (c *chatServiceClient) GetUnreadCount(ctx context.Context, in *GetUnreadCountRequest, opts ...grpc.CallOption) (*GetUnreadCountResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(GetUnreadCountResponse)
+	err := c.cc.Invoke(ctx, ChatService_GetUnreadCount_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 // ChatServiceServer is the server API for ChatService service.
 // All implementations must embed UnimplementedChatServiceServer
 // for forward compatibility.
@@ -1162,6 +1174,7 @@ type ChatServiceServer interface {
 	SubscribeNotifications(*SubscribeNotificationsRequest, grpc.ServerStreamingServer[ServerNotification]) error
 	GetNotificationHistory(context.Context, *GetNotificationHistoryRequest) (*GetNotificationHistoryResponse, error)
 	MarkNotificationsRead(context.Context, *MarkNotificationReadRequest) (*MarkNotificationReadResponse, error)
+	GetUnreadCount(context.Context, *GetUnreadCountRequest) (*GetUnreadCountResponse, error)
 	mustEmbedUnimplementedChatServiceServer()
 }
 
@@ -1417,6 +1430,9 @@ func (UnimplementedChatServiceServer) GetNotificationHistory(context.Context, *G
 }
 func (UnimplementedChatServiceServer) MarkNotificationsRead(context.Context, *MarkNotificationReadRequest) (*MarkNotificationReadResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method MarkNotificationsRead not implemented")
+}
+func (UnimplementedChatServiceServer) GetUnreadCount(context.Context, *GetUnreadCountRequest) (*GetUnreadCountResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetUnreadCount not implemented")
 }
 func (UnimplementedChatServiceServer) mustEmbedUnimplementedChatServiceServer() {}
 func (UnimplementedChatServiceServer) testEmbeddedByValue()                     {}
@@ -2854,6 +2870,24 @@ func _ChatService_MarkNotificationsRead_Handler(srv interface{}, ctx context.Con
 	return interceptor(ctx, in, info, handler)
 }
 
+func _ChatService_GetUnreadCount_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(GetUnreadCountRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(ChatServiceServer).GetUnreadCount(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: ChatService_GetUnreadCount_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(ChatServiceServer).GetUnreadCount(ctx, req.(*GetUnreadCountRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 // ChatService_ServiceDesc is the grpc.ServiceDesc for ChatService service.
 // It's only intended for direct use with grpc.RegisterService,
 // and not to be introspected or modified (even as a copy)
@@ -3160,6 +3194,10 @@ var ChatService_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "MarkNotificationsRead",
 			Handler:    _ChatService_MarkNotificationsRead_Handler,
+		},
+		{
+			MethodName: "GetUnreadCount",
+			Handler:    _ChatService_GetUnreadCount_Handler,
 		},
 	},
 	Streams: []grpc.StreamDesc{